requirements.txt
streamlit>=1.32.0
snowflake-connector-python>=3.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
